requirements.txt
aiohttp
python-socketio[asyncio]==4.6.1
fastapi
uvicorn